+++ src/shapes/IShape.h
@@ -74,7 +74,7 @@ public:
     /**
      * @brief Writes a float description of the shape into the given array. Max length DESC_LEN!
      */
-    virtual void writeDescription (float *desc_array) const;
+    virtual void writeDescription (int *desc_array) const;
 #endif
